requirements.txt
django==3.2.16
djangorestframework==3.14.0
django-cors-headers